WIP Infested Models & Tweaks
- Updated the models and textures for infested crawlers (and bashers) to be more complete.
- Fixed some various script syntax errors.
- Moved the artifact for Gas Turbine Generator Mk2 from A3L2 to A3L7.
- Slowed down Golf's factories on A4L3.
- Significantly weakened Foxtrot's factories on A4L3, and slightly weakened Golf's while removing a few Howitzer/RR emplacements.
- Collective kill teams no longer stop spawning in A4L5 on INSANE.
- Fixed Collective base trucks not spawning until stage 3 on A4L6.
- Fixed the Collective "de-aggro" zone not resetting properly on A4L6.
- Dimmed the stage 3 lightning effects slightly.

+++ script/campaign/A4L5.js
@@ -392,7 +392,11 @@ function groundAssault4()
 	activateGroundBlip(6);
 	activateGroundBlip(7);
 
-	killTeamsEnabled = false; // Stop spawning Collective kill teams
+
+	if (difficulty < INSANE)
+	{
+		killTeamsEnabled = false; // Stop spawning Collective kill teams
+	}
 	
 	playSound(cam_sounds.enemyUnitDetected);
